k8s/api-server/app/requirements.txt
fastapi==0.75.0
pydantic==1.9.0
uvicorn==0.17.6
nats-py==2.1.0